@@ -1,4 +1,4 @@
-// swift-tools-version: 5.6
+// swift-tools-version: 5.8
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -6,7 +6,7 @@ import PackageDescription
 let package = Package(
     name: "UserInterface",
     platforms: [
-        .macOS(.v12),
+        .macOS(.v13),
         .macCatalyst(.v15),
         .iOS(.v15)
     ],
