@@ -7,6 +7,19 @@
 
 import SwiftUI
 
+extension View {
+    @ViewBuilder
+    func conditionalScrollBounce() -> some View {
+        if #available(macOS 13.3, *) {
+            self
+                .scrollBounceBehavior(.basedOnSize)
+        }
+        else {
+            self
+        }
+    }
+}
+
 struct LaunchpadView: View {
     @StateObject var launchpadController: LaunchpadController
     @State private var repos: [LaunchpadRepo] = []
@@ -14,12 +27,16 @@ struct LaunchpadView: View {
     @State private var selectedView: Int = 0
 
     var body: some View {
-        HStack() {
-            ForEach(repos, id: \.id) { repo in
-                LaunchpadItem(repo: repo)
+        ScrollView(.horizontal) {
+            HStack() {
+                ForEach(repos.reversed(), id: \.id) { repo in
+                    LaunchpadItem(repo: repo)
+                }
+                Spacer()
             }
-            Spacer()
         }
+        .scrollIndicators(.hidden)
+        .conditionalScrollBounce()
         .padding(.leading)
         .onReceive(launchpadController.$contributedRepos.$items, perform: {
             // We can even create complex pipelines, for example filtering all notes bigger than a tweet
