@@ -1,6 +1,6 @@
 //
 //  AccountSettingsView.swift
-//  
+//
 //
 //  Created by Stef Kors on 21/07/2022.
 //
@@ -9,35 +9,39 @@ import SwiftUI
 
 public struct AccountSettingsView: View {
     @EnvironmentObject public var model: NetworkManager
-    
-    public var body: some View {
-        VStack(alignment: .leading) {
-            GroupBox {
-                HStack {
-                    Text("GitLab Token")
-                    VStack(alignment: .leading) {
-                        HStack {
-                            TextField(
-                                "Enter token here...",
-                                text: NetworkManager.$apiToken,
-                                onCommit: {
-                                    // make API call with token.
-                                    Task {
-                                        await model.fetch()
-                                    }
-                                })
-                            .textFieldStyle(RoundedBorderTextFieldStyle())
-                            Button("Clear API Token", action: {NetworkManager.apiToken = ""})
-                        }
 
-                        Text("Create a read-only GitLab [access-token](https://gitlab.com/-/profile/personal_access_tokens) that the app can")
-                            .foregroundColor(.secondary)
-                        Text("use to query the API with.")
-                            .foregroundColor(.secondary)
+    public var body: some View {
+        Form {
+            HStack(alignment: .top) {
+                Text("GitLab Token")
+                VStack(alignment: .leading) {
+                    HStack {
+                        TextField("GitLab Token", text: NetworkManager.$apiToken, prompt: Text("Enter token here..."))
+                            .textFieldStyle(RoundedBorderTextFieldStyle())
+                            .labelsHidden()
+                            .onSubmit(of: .text) {
+                                Task {
+                                    await model.fetch()
+                                }
+                            }
+
+                        Button("Update", action: {
+                            Task {
+                                await model.fetch()
+                            }
+                        })
+
+                        Button("Clear", action: {NetworkManager.apiToken = ""})
                     }
-                }.padding()
+
+                    Text("Create a read-only GitLab [access-token](https://gitlab.com/-/profile/personal_access_tokens) that the app can")
+                        .foregroundColor(.secondary)
+                    Text("use to query the API with.")
+                        .foregroundColor(.secondary)
+                }
             }
-        }.padding()
+        }
+        .formStyle(.grouped)
     }
 }
 
